% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq.R
\name{run_iq}
\alias{run_iq}
\title{Performs Installation Qualification}
\usage{
run_iq(location)
}
\arguments{
\item{location}{The path to the desired output directory.  The IQ
reports and any associated files will be placed in this directory.}
}
\value{
The path to the output directory.
}
\description{
The \code{run_iq} function executes an
installation qualification (IQ)
on the currently installed \strong{sassy} packages, and generates a report
on the results. The IQ ensures that the files in the \strong{sassy} packages
are correct, and have not been altered.
The results of the IQ will be placed in the supplied location.
}
\examples{
# Create a temp directory
tmp <- tempdir()

# Run the Installation Qualification
run_iq(tmp)
}
