% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oq.R
\name{run_oq}
\alias{run_oq}
\title{Performs Operational Qualification}
\usage{
run_oq(location)
}
\arguments{
\item{location}{The path to the desired output directory.  The IQ
reports and any other associated files will be placed in this directory.
Location should be specified as a directory only.  The file name
will be generated by the function.}
}
\value{
The path to the output directory.
}
\description{
The \code{run_iq} function executes an
operational qualification (OQ)
on the currently installed \strong{sassy} packages and generates a report
on the results.  The OQ ensures that the \strong{sassy} packages are working
as expected.  The results will be placed in the supplied location.
}
\examples{
# Create a temp directory
tmp <- tempdir()

# Run the Operational Qualification
run_oq(tmp)
}
