% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sassy.R
\docType{package}
\name{sassy}
\alias{sassy}
\title{A Collection of Packages Inspired by SAS®}
\description{
The \strong{sassy} package is a meta-package that
installs a set of packages
inspired by SAS® software.  These packages include the ability to create
a traceable log, create data libraries and format catalogs, and includes
reporting capabilities similar to those found in SAS®.  This package was
written independently, and the authors have no association with or
endorsement of SAS Institute.
}
\section{Packages Included}{

The packages included in the \strong{sassy} meta-package are
as follows:
\itemize{
  \item {\code{\link[logr]{logr}}:}{ Write a traceable log.}
  \item {\code{\link[fmtr]{fmtr}}:}{ Create a format catalog
and apply formats to a data frame or vector.}
}
}

